added responsive layout and refactored chart to another file
